requirements.txt
astroid==2.3.3
cycler==0.10.0
dj-database-url==0.5.0
Django==2.2.5
django-extensions==2.2.5
django-storages==1.7.2
django-heroku==0.3.1
docutils==0.15.2
et-xmlfile==1.0.1
gunicorn==20.0.0
isort==4.3.21
jdcal==1.4.1
jmespath==0.9.4
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
matplotlib==3.1.2
mccabe==0.6.1
numpy==1.17.4
openpyxl==3.0.2
Pillow==6.1.0
psycopg2==2.8.3
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
pyparsing==2.4.5
python-dateutil==2.8.0
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.7
whitenoise==4.1.4
wrapt==1.11.2
django-import-export==2.0.1
xhtml2pdf==0.2.4 
